Authenticate snippets reorder request on backend
