+++ app/Http/Controllers/SnippetController.php
@@ -136,4 +136,22 @@ class SnippetController extends Controller
         //
     }
 
+    /**
+     * Reorder snippets
+     *
+     * @param  \Illuminate\Http\Request  $request
+     */
+    public function updateOrder(Request $request)
+    {
+        $user = \Auth::user();
+
+        $idsOrders = $request->input('idsOrders');
+        foreach ($idsOrders as $idOrder) {
+            $snippet = Snippet::where('id', $idOrder['id'])
+                ->where('user_id', $user->id)
+                ->firstOrFail();
+            $snippet->order = $idOrder['order'];
+            $snippet->save();
+        }
+    }
 }
