+++ routes/api.php
@@ -1,7 +1,6 @@
 <?php
 
 use App\Http\Controllers\API\EntryController;
-use App\Http\Controllers\API\SnippetController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -19,9 +18,7 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->group(function () {
     Route::get('/entries/{id}', [EntryController::class, 'get'])->name('api.entries.id');
     Route::get('/entries/more/{id}',[EntryController::class, 'moreEntries'])->name('api.entries.more');
-    // Route::post('/snippets/update-order', [SnippetController::class, 'updateOrder'])->name('api.snippets.update-order');
 });
 
-Route::post('/snippets/update-order', [SnippetController::class, 'updateOrder'])->name('api.snippets.update-order');
 // one-off
 // Route::middleware('auth:sanctum')->get('/entries/{id}', [EntryController::class, 'get'])->name('api.entries.id');
